feat: extracting pageBuilder to its own function so that is reusable in multiple pages
